+++ commands/src/main/java/net/apartium/cocoabeans/commands/CommandManager.java
@@ -16,7 +16,6 @@ import net.apartium.cocoabeans.commands.exception.ExceptionArgumentMapper;
 import net.apartium.cocoabeans.commands.exception.HandleExceptionVariant;
 import net.apartium.cocoabeans.commands.exception.UnknownCommandResponse;
 import net.apartium.cocoabeans.commands.parsers.*;
-import net.apartium.cocoabeans.commands.parsers.factory.ParserFactory;
 import net.apartium.cocoabeans.commands.requirements.*;
 import org.jetbrains.annotations.ApiStatus;