+++ commands/src/main/java/net/apartium/cocoabeans/commands/parsers/WithParsers.java
@@ -20,6 +20,7 @@ import java.lang.annotation.Target;
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE, ElementType.METHOD})
+@CommandParserFactory(value = WithParserFactory.class, scope = Scope.ALL)
 public @interface WithParsers {
 
     /**