+++ commands/src/main/java/net/apartium/cocoabeans/commands/parsers/factory/IntRangeParserFactory.java
@@ -15,16 +15,27 @@ import net.apartium.cocoabeans.commands.CommandNode;
 import net.apartium.cocoabeans.commands.CommandProcessingContext;
 import net.apartium.cocoabeans.commands.parsers.ArgumentParser;
 import net.apartium.cocoabeans.commands.parsers.IntRangeParser;
+import net.apartium.cocoabeans.commands.parsers.ParserFactory;
+import net.apartium.cocoabeans.commands.parsers.Scope;
+import org.jetbrains.annotations.ApiStatus;
+import org.jetbrains.annotations.NotNull;
 
+import java.lang.annotation.Annotation;
+import java.lang.reflect.GenericDeclaration;
+import java.lang.reflect.Method;
 import java.util.*;
 import java.util.stream.Collectors;
 
+/**
+ * @hidden
+ */
+@ApiStatus.Internal
 public class IntRangeParserFactory implements ParserFactory {
 
     @Override
-    public ArgumentParser<?> getArgumentParser(CommandNode commandNode, Object obj) {
-        if (!(obj instanceof IntRangeParser intRangeParser))
-            return null;
+    public @NotNull Collection<ParserResult> getArgumentParser(CommandNode commandNode, Annotation annoation, GenericDeclaration obj) {
+        if (!(annoation instanceof IntRangeParser intRangeParser))
+            return Collections.emptyList();
 
         if (intRangeParser.from() > intRangeParser.to())
             throw new RuntimeException("from must be smaller than to");
@@ -34,7 +45,10 @@ public class IntRangeParserFactory implements ParserFactory {
         if ((intRangeParser.to() - intRangeParser.from()) <= intRangeParser.step())
             throw new RuntimeException("step must be smaller than to - from");
 
-        return new IntRangeParserImpl(intRangeParser.from(), intRangeParser.to(), intRangeParser.step(), intRangeParser.keyword(), intRangeParser.priority());
+        return List.of(new ParserResult(
+                new IntRangeParserImpl(intRangeParser.from(), intRangeParser.to(), intRangeParser.step(), intRangeParser.keyword(), intRangeParser.priority()),
+                obj instanceof Method ? Scope.VARIANT : Scope.CLASS
+        ));
     }
 
     private class IntRangeParserImpl extends ArgumentParser<Integer> {