+++ spigot/src/main/java/net/apartium/cocoabeans/spigot/inventory/ItemUtilsHelpers.java
@@ -1,0 +1,14 @@
+package net.apartium.cocoabeans.spigot.inventory;
+
+import org.bukkit.Material;
+import org.jetbrains.annotations.ApiStatus;
+
+/**
+ * @hidden
+ */
+@ApiStatus.Internal
+public interface ItemUtilsHelpers {
+
+    String name(Material material);
+
+}